# fastapi-web-server/docker-code-gen/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "UUDEX API",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/subject/{subject_id}": {
            "get": {
                "tags": [
                    "subjects"
                ],
                "summary": "Get Subject By Id",
                "operationId": "get_subject_by_id_subject__subject_id__get",
                "parameters": [
                    {
                        "name": "subject_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Subject Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subject"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/subjects/": {
            "get": {
                "tags": [
                    "subjects"
                ],
                "summary": "Get All Subjects",
                "operationId": "get_all_subjects_subjects__get",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticatedUser"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Subject"
                                    },
                                    "type": "array",
                                    "title": "Response Get All Subjects Subjects  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthenticatedUser": {
                "properties": {
                    "endpoint_id": {
                        "type": "integer",
                        "title": "Endpoint Id"
                    },
                    "endpoint_uuid": {
                        "type": "string",
                        "title": "Endpoint Uuid"
                    },
                    "endpoint_user_name": {
                        "type": "string",
                        "title": "Endpoint User Name"
                    }
                },
                "type": "object",
                "required": [
                    "endpoint_id",
                    "endpoint_uuid",
                    "endpoint_user_name"
                ],
                "title": "AuthenticatedUser"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Subject": {
                "properties": {
                    "create_datetime": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Create Datetime"
                    },
                    "subject_uuid": {
                        "type": "string",
                        "title": "Subject Uuid"
                    },
                    "subject_name": {
                        "type": "string",
                        "title": "Subject Name"
                    },
                    "dataset_instance_key": {
                        "type": "string",
                        "title": "Dataset Instance Key"
                    },
                    "subscription_type": {
                        "type": "string",
                        "title": "Subscription Type"
                    },
                    "fulfillment_types_available": {
                        "type": "string",
                        "title": "Fulfillment Types Available"
                    },
                    "full_queue_behavior": {
                        "type": "string",
                        "title": "Full Queue Behavior"
                    },
                    "max_queue_size_kb": {
                        "type": "integer",
                        "title": "Max Queue Size Kb"
                    },
                    "max_message_count": {
                        "type": "integer",
                        "title": "Max Message Count"
                    },
                    "priority": {
                        "type": "integer",
                        "title": "Priority"
                    },
                    "backing_exchange_name": {
                        "type": "string",
                        "title": "Backing Exchange Name"
                    },
                    "subject_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Subject Id"
                    },
                    "owner_participant_id": {
                        "type": "integer",
                        "title": "Owner Participant Id"
                    },
                    "dataset_definition_id": {
                        "type": "integer",
                        "title": "Dataset Definition Id"
                    }
                },
                "type": "object",
                "required": [
                    "subject_uuid",
                    "subject_name",
                    "dataset_instance_key",
                    "subscription_type",
                    "fulfillment_types_available",
                    "full_queue_behavior",
                    "max_queue_size_kb",
                    "max_message_count",
                    "priority",
                    "backing_exchange_name",
                    "owner_participant_id",
                    "dataset_definition_id"
                ],
                "title": "Subject"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    },
...
}